Move TriggeredCommandFormTests.
